requirements.txt
streamlit
pandas
openpyxl
PyMuPDF
Pillow
pdf2docx
fpdf2